requirements.txt
instructor>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
